carga só é embarcada se navio tiver capacidade

+++ domain/supervisor.rb
@@ -2,9 +2,11 @@ require File.join(File.dirname(File.expand_path(__FILE__)), '..', 'config', 'ar_
 
 class Supervisor
 	def embarcar (carga, navio)
-		atributos_de_carga = carga.instance_values["attributes"] 
-		atributos_de_carga["navio_id"] = navio.id
-		carga.destroy
-		navio.carga_embarcadas << CargaEmbarcada.create!(atributos_de_carga) 
+		if carga.peso <= navio.capacidade
+			atributos_de_carga = carga.instance_values["attributes"] 
+			atributos_de_carga["navio_id"] = navio.id
+			carga.destroy
+			navio.carga_embarcadas << CargaEmbarcada.create!(atributos_de_carga) 
+		end
 	end
 end
